Alterações para padronização

+++ src/main/java/br/com/autochef/AutoChef/dto/restaurant/DetailsRestaurantDto.java
@@ -2,7 +2,9 @@ package br.com.autochef.AutoChef.dto.restaurant;
 
 import br.com.autochef.AutoChef.model.RestaurantModel;
 
-public record DetailsRestaurantDto(String name, String imageUrl, Float longitude, Float latitude) {
+import java.math.BigDecimal;
+
+public record DetailsRestaurantDto(String name, String imageUrl, BigDecimal longitude, BigDecimal latitude) {
 
     public DetailsRestaurantDto(RestaurantModel rest){
         this(rest.getName(), rest.getImageUrl(), rest.getLatitude(), rest.getLongitude());
