+++ src/main/java/br/com/autochef/AutoChef/dto/restaurant/UpdateRestaurantDto.java
@@ -2,8 +2,10 @@ package br.com.autochef.AutoChef.dto.restaurant;
 
 import br.com.autochef.AutoChef.model.RestaurantModel;
 
+import java.math.BigDecimal;
+
 public record UpdateRestaurantDto(String name, String email, String password, String imageUrl,
-                                  Float longitude, Float latitude) {
+                                  BigDecimal longitude, BigDecimal latitude) {
 
     public UpdateRestaurantDto (RestaurantModel rest){
         this(rest.getName(), rest.getEmail(), rest.getPassword(), rest.getImageUrl(),
